Introduce RoomListRoomSummaryFactory and remove RoomListRoomSummaryPlaceholders.
