+++ features/roomlist/impl/src/main/kotlin/io/element/android/features/roomlist/impl/RoomListStateProvider.kt
@@ -19,7 +19,6 @@ package io.element.android.features.roomlist.impl
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import io.element.android.features.leaveroom.api.aLeaveRoomState
 import io.element.android.features.roomlist.impl.model.RoomListRoomSummary
-import io.element.android.features.roomlist.impl.model.RoomListRoomSummaryPlaceholders
 import io.element.android.features.roomlist.impl.model.aRoomListRoomSummary
 import io.element.android.libraries.designsystem.components.avatar.AvatarData
 import io.element.android.libraries.designsystem.components.avatar.AvatarSize
@@ -88,7 +87,13 @@ internal fun aRoomListRoomSummaryList(): ImmutableList<RoomListRoomSummary> {
             avatarData = AvatarData("!id", "Z", size = AvatarSize.RoomListItem),
             id = "!roomId2:domain",
         ),
-        RoomListRoomSummaryPlaceholders.create("!roomId3:domain"),
-        RoomListRoomSummaryPlaceholders.create("!roomId4:domain"),
+        aRoomListRoomSummary(
+            id = "!roomId3:domain",
+            isPlaceholder = true,
+        ),
+        aRoomListRoomSummary(
+            id = "!roomId4:domain",
+            isPlaceholder = true,
+        ),
     )
 }
